+++ lib/metapede_web/templates/page_views/time_period_show.html.leex
@@ -20,6 +20,7 @@
 		<%= live_component @socket,
 			MetapedeWeb.LiveComponents.TimePeriod.SubPeriodComponent,
 			time_period: @time_period,
+			new_topic: @new_topic,
 			live_action: @live_action
 		%>
 	</div>
